requirements.txt
gymnasium==0.28.1
numpy==1.24.3
stable-baselines3==2.1.0
torch==2.0.1
pygame==2.5.0
PyOpenGL==3.1.6
matplotlib==3.7.1

